+++ crates/wdl-ast/CHANGELOG.md
@@ -5,15 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.2.0 - 5-31-2024
+## Unreleased
 
-* Fix ignoring comments in expressions ([#23](https://github.com/stjude-rust-labs/wdl/pull/23)).
+### Fixed
+
+* Fixed the experimental parser validation to check negative numbers in
+  metadata sections ([#66](https://github.com/stjude-rust-labs/wdl/pull/66)).
 
 ### Added
 
 * Added validation rules for the experimental AST ([#65](https://github.com/stjude-rust-labs/wdl/pull/65)).
 * Added a new experimental AST for the experimental parser; this implementation
   is currently feature-gated behind the `experimental` feature ([#64](https://github.com/stjude-rust-labs/wdl/pull/64)).
+
+## 0.2.0 - 5-31-2024
+
+* Fix ignoring comments in expressions ([#23](https://github.com/stjude-rust-labs/wdl/pull/23)).
 
 ### Changed